Update refresh script
This commit changes refresh script to append items instead of replacing
the whole channel. It adds the 'append' config variable to establish the
number of items to append. The 'limit' config parameter sets the maximum
number of items permitted in a channel, old entries are deleted to match
the limit.

Signed-off-by: David Cassany <dcassany@suse.com>